+++ docs/changelog.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+* Change ``reswap()`` type hint for ``method`` to ``str``.
+
+  Thanks to Dan Jacob for the report in `Issue #421 <https://github.com/adamchainz/django-htmx/issues/421>`__ and fix in `PR #422 <https://github.com/adamchainz/django-htmx/pull/422>`__.
+
 1.17.2 (2023-11-16)
 -------------------